infrastructure/requirements-dev.txt
# Development dependencies for infrastructure testing
pytest>=7.4.0
aws-cdk-lib>=2.100.0
constructs>=10.0.0
